requirements.txt
discord.py==2.3.2
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
